+++ test/event/event_handler_state_test.exs
@@ -1,7 +1,6 @@
 defmodule Commanded.Event.EventHandlerStateTest do
-  use ExUnit.Case
+  use Commanded.MockEventStoreCase
 
-  alias Commanded.DefaultApp
   alias Commanded.Event.StatefulEventHandler
   alias Commanded.Helpers.EventFactory
 
@@ -11,15 +10,10 @@ defmodule Commanded.Event.EventHandlerStateTest do
   end
 
   describe "event handler state" do
-    setup do
-      start_supervised!(DefaultApp)
-      :ok
-    end
-
     test "initially set in `init/1` function" do
       handler = start_supervised!(StatefulEventHandler)
 
-      event = %AnEvent{reply_to: self(), update_state?: true}
+      event = %AnEvent{reply_to: reply_to(), update_state?: true}
       send_events_to_handler(handler, [event])
 
       assert_receive {:event, ^event, metadata}
@@ -29,7 +23,7 @@ defmodule Commanded.Event.EventHandlerStateTest do
     test "initially set as runtime option" do
       handler = start_supervised!({StatefulEventHandler, state: 1})
 
-      event = %AnEvent{reply_to: self(), update_state?: true}
+      event = %AnEvent{reply_to: reply_to(), update_state?: true}
       send_events_to_handler(handler, [event])
 
       assert_receive {:event, ^event, metadata}
@@ -39,8 +33,8 @@ defmodule Commanded.Event.EventHandlerStateTest do
     test "updated by returning `{:ok, new_state}` from `handle/2` function" do
       handler = start_supervised!(StatefulEventHandler)
 
-      event1 = %AnEvent{reply_to: self(), update_state?: true}
-      event2 = %AnEvent{reply_to: self(), update_state?: true}
+      event1 = %AnEvent{reply_to: reply_to(), update_state?: true}
+      event2 = %AnEvent{reply_to: reply_to(), update_state?: true}
       send_events_to_handler(handler, [event1, event2])
 
       assert_receive {:event, ^event1, metadata}
@@ -53,8 +47,8 @@ defmodule Commanded.Event.EventHandlerStateTest do
     test "not updated when returning `:ok` from `handle/2` function" do
       handler = start_supervised!(StatefulEventHandler)
 
-      event1 = %AnEvent{reply_to: self(), update_state?: false}
-      event2 = %AnEvent{reply_to: self(), update_state?: false}
+      event1 = %AnEvent{reply_to: reply_to(), update_state?: false}
+      event2 = %AnEvent{reply_to: reply_to(), update_state?: false}
       send_events_to_handler(handler, [event1, event2])
 
       assert_receive {:event, ^event1, metadata}
@@ -68,8 +62,8 @@ defmodule Commanded.Event.EventHandlerStateTest do
       opts = [state: 10]
       handler = start_supervised!({StatefulEventHandler, opts})
 
-      event1 = %AnEvent{reply_to: self(), update_state?: true}
-      event2 = %AnEvent{reply_to: self(), update_state?: true}
+      event1 = %AnEvent{reply_to: reply_to(), update_state?: true}
+      event2 = %AnEvent{reply_to: reply_to(), update_state?: true}
       send_events_to_handler(handler, [event1, event2])
 
       assert_receive {:event, ^event1, metadata}
@@ -84,13 +78,15 @@ defmodule Commanded.Event.EventHandlerStateTest do
 
       handler = start_supervised!({StatefulEventHandler, opts})
 
-      event3 = %AnEvent{reply_to: self(), update_state?: true}
+      event3 = %AnEvent{reply_to: reply_to(), update_state?: true}
       send_events_to_handler(handler, [event3], 3)
 
       assert_receive {:event, ^event3, metadata}
       assert match?(%{state: 10}, metadata)
     end
   end
+
+  defp reply_to, do: :erlang.pid_to_list(self())
 
   defp send_events_to_handler(handler, events, initial_event_number \\ 1) do
     recorded_events = EventFactory.map_to_recorded_events(events, initial_event_number)