+++ test/event/support/echo_handler.ex
@@ -1,6 +1,6 @@
 defmodule Commanded.Event.EchoHandler do
   use Commanded.Event.Handler,
-    application: Commanded.DefaultApp,
+    application: Commanded.MockedApp,
     name: __MODULE__
 
   alias Commanded.Event.ReplyEvent
@@ -8,6 +8,8 @@ defmodule Commanded.Event.EchoHandler do
   @impl Commanded.Event.Handler
   def handle(%ReplyEvent{} = event, metadata) do
     %ReplyEvent{reply_to: reply_to} = event
+
+    reply_to = :erlang.list_to_pid(reply_to)
 
     send(reply_to, {:event, self(), event, metadata})