+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Allow a process manager to stop after dispatching a command ([#460](https://github.com/commanded/commanded/pull/460)).
 - Replace `use Mix.Config` with `import Config` in config files ([#467](https://github.com/commanded/commanded/pull/467)).
+- Event handler concurrency ([#486](https://github.com/commanded/commanded/pull/486)).
 
 ## v1.3.1